requirements.txt
evalplus[vllm] @ git+https://github.com/evalplus/evalplus.git
openai>=1.30.0
azure-identity
azure-mgmt-cognitiveservices
azure-ai-projects>=2.0.0b1
azure-ai-evaluation>=1.13.0
python-dotenv
evalplus>=0.3.0
radon>=6.0.0
datasets>=2.18.0
tqdm
scipy
numpy
